+++ include/device/LinearOp.hpp
@@ -13,7 +13,7 @@ namespace SE{
 
 //memory managament
 template<typename DATATYPE, DEVICETYPE devicetype =DEVICETYPE::BASE>
-DATATYPE* malloc(const size_t size){
+DATATYPE* malloc(const int size){
     return static_cast<DATATYPE*>(std::malloc(size * sizeof(DATATYPE)));
 }
 
@@ -23,7 +23,7 @@ void free(void* ptr){
 }
 
 template<typename DATATYPE, DEVICETYPE devicetype=DEVICETYPE::BASE>
-void memcpy(DATATYPE* dest, const DATATYPE* source, size_t size, COPYTYPE copy_type=COPYTYPE::NONE){
+void memcpy(DATATYPE* dest, const DATATYPE* source, int size, COPYTYPE copy_type=COPYTYPE::NONE){
     std::cout << typeid(copy_type).name() <<"\t"<< typeid(COPYTYPE::NONE).name() <<std::endl;
     std::cout << (int) copy_type <<"\t" << (int) COPYTYPE::NONE <<std::endl;
     assert(COPYTYPE::NONE==(COPYTYPE) copy_type );
@@ -31,32 +31,32 @@ void memcpy(DATATYPE* dest, const DATATYPE* source, size_t size, COPYTYPE copy_t
 }
 
 template<typename DATATYPE, DEVICETYPE devicetype=DEVICETYPE::BASE>
-void memset(DATATYPE* dest, int value, size_t size){
+void memset(DATATYPE* dest, int value, int size){
     std::memset(dest, value, size * sizeof(DATATYPE));
 }
 
 
 //x = a * x
 template <typename DATATYPE, DEVICETYPE devicetype>
-void scal(const size_t n, const DATATYPE alpha, DATATYPE *x, const size_t incx);
+void scal(const int n, const DATATYPE alpha, DATATYPE *x, const int incx);
 
 //a * x + y
 template <typename DATATYPE, DEVICETYPE devicetype>
-void axpy(const size_t n, const DATATYPE a, const DATATYPE *x, const size_t incx, DATATYPE *y, const size_t incy);
+void axpy(const int n, const DATATYPE a, const DATATYPE *x, const int incx, DATATYPE *y, const int incy);
 
 //Euclidean norm, ||x||
 template <typename DATATYPE, DEVICETYPE devicetype>
-DATATYPE nrm2(const size_t n, const DATATYPE *x, const size_t incx);
+DATATYPE nrm2(const int n, const DATATYPE *x, const int incx);
 
 //y (i*N+incy) = x (i*M+incx) 
 template <typename DATATYPE, DEVICETYPE devicetype>
-void copy(const size_t n, const DATATYPE *x, const size_t incx, DATATYPE *y, const size_t incy);
+void copy(const int n, const DATATYPE *x, const int incx, DATATYPE *y, const int incy);
 
 //alpha * A * x + beta * y
 template <typename DATATYPE, DEVICETYPE devicetype>
-void gemv(const ORDERTYPE layout, const TRANSTYPE transa, const size_t m, const size_t n, const DATATYPE alpha,
-          const DATATYPE *a, const size_t lda, const DATATYPE *x, const size_t incx,
-          const DATATYPE beta, DATATYPE *y, const size_t incy);
+void gemv(const ORDERTYPE layout, const TRANSTYPE transa, const int m, const int n, const DATATYPE alpha,
+          const DATATYPE *a, const int lda, const DATATYPE *x, const int incx,
+          const DATATYPE beta, DATATYPE *y, const int incy);
 
 //alpha * A * x + b * y
 //void coomv;
@@ -68,17 +68,17 @@ void gemv(const ORDERTYPE layout, const TRANSTYPE transa, const size_t m, const 
 //alpha * A * B + beta * C, A : m by k, B : k by n, C : m by n
 //output : C
 template <typename DATATYPE, DEVICETYPE devicetype>
-void gemm(const ORDERTYPE layout, const TRANSTYPE transa, const TRANSTYPE transb, const size_t m, const size_t n, const size_t k,
-          const DATATYPE alpha, const DATATYPE *a, const size_t lda,
-          const DATATYPE *b, const size_t ldb, const DATATYPE beta,
-          DATATYPE *c, const size_t ldc);
+void gemm(const ORDERTYPE layout, const TRANSTYPE transa, const TRANSTYPE transb, const int m, const int n, const int k,
+          const DATATYPE alpha, const DATATYPE *a, const int lda,
+          const DATATYPE *b, const int ldb, const DATATYPE beta,
+          DATATYPE *c, const int ldc);
 
 //Blas-like extensions
-//void mkl_domatcopy (char ordering, char trans, size_t rows, size_t cols,
-//                   const double alpha, const double * A, size_t lda, double * B, size_t ldb);
+//void mkl_domatcopy (char ordering, char trans, int rows, int cols,
+//                   const double alpha, const double * A, int lda, double * B, int ldb);
 template <typename DATATYPE, DEVICETYPE devicetype>
-void omatcopy(const ORDERTYPE layout, const TRANSTYPE trans, size_t rows, size_t cols,
-              const double alpha, DATATYPE *a, size_t lda, DATATYPE *b, size_t ldb);
+void omatcopy(const ORDERTYPE layout, const TRANSTYPE trans, int rows, int cols,
+              const double alpha, DATATYPE *a, int lda, DATATYPE *b, int ldb);
 
 
 //LAPACK
@@ -87,7 +87,7 @@ void omatcopy(const ORDERTYPE layout, const TRANSTYPE trans, size_t rows, size_t
 //for the substentiall tall matrix, see geqr
 //If we don't need to directly apply q matrix and just need to multiply with othter matrix, see ormqr
 template <typename DATATYPE, DEVICETYPE devicetype>
-int geqrf(const ORDERTYPE layout, size_t m, size_t n, DATATYPE* a, size_t lda, DATATYPE* tau);
+int geqrf(const ORDERTYPE layout, int m, int n, DATATYPE* a, int lda, DATATYPE* tau);
 
 //generate real orthogonal matrix Q from geqrf
 //LAPACKE_dorgqr(LAPACK_COL_MAJOR, vector_size, number_of_vectors, number_of_vectors, eigvec, vector_size, tau);
@@ -95,7 +95,7 @@ int geqrf(const ORDERTYPE layout, size_t m, size_t n, DATATYPE* a, size_t lda, D
 //                           lapack_int k, double* a, lapack_int lda,
 //                           const double* tau );
 template <typename DATATYPE, DEVICETYPE devicetype>
-int orgqr(const ORDERTYPE layout, size_t m, size_t n, DATATYPE* a, size_t lda, DATATYPE* tau);
+int orgqr(const ORDERTYPE layout, int m, int n, DATATYPE* a, int lda, DATATYPE* tau);
 
 //ormqr
 
@@ -105,10 +105,10 @@ int orgqr(const ORDERTYPE layout, size_t m, size_t n, DATATYPE* a, size_t lda, D
                           //double* wi, double* vl, lapack_int ldvl, double* vr,
 //                          lapack_int ldvr );
 template <typename DATATYPE, DEVICETYPE devicetype>
-int geev(const ORDERTYPE layout, char jobvl, char jobvr, const size_t n, DATATYPE* a, const size_t lda,
-          DATATYPE* wr, DATATYPE* wi, DATATYPE* vl, const size_t ldvl, DATATYPE* vr, const size_t ldvr);
+int geev(const ORDERTYPE layout, char jobvl, char jobvr, const int n, DATATYPE* a, const int lda,
+          DATATYPE* wr, DATATYPE* wi, DATATYPE* vl, const int ldvl, DATATYPE* vr, const int ldvr);
 
 
 template <typename DATATYPE, DEVICETYPE devicetype>
-int syev(const ORDERTYPE layout, char jobz, char uplo, const size_t n, double* a, const size_t lda, double* w);
+int syev(const ORDERTYPE layout, char jobz, char uplo, const int n, double* a, const int lda, double* w);
 }
